@@ -1,97 +1,101 @@
 {
-  "name": "rmwc-main",
-  "author": "James Friedman",
-  "version": "6.1.1",
-  "private": true,
-  "homepage": "https://rmwc.io",
-  "description": "RMWC is a React UI Kit built on Google's official Material Components Web library.",
-  "license": "MIT",
-  "sideEffects": true,
-  "main": "index.js",
-  "types": "index.d.ts",
-  "peerDependencies": {
-    "react": ">=16.8.x",
-    "react-dom": ">=16.8.x"
-  },
-  "scripts": {
-    "install": "lerna bootstrap --force-local --hoist",
-    "start": "export NODE_OPTIONS='--max-old-space-size=4096' && react-app-rewired start",
-    "build:docsite": "export NODE_OPTIONS='--max-old-space-size=4096' && react-app-rewired build && rm -R -f docs && mv build docs",
-    "build:docs": "npm run clean && npm run docgen && npm run build:docsite",
-    "build:code": "npm run clean && node scripts/build-code.js",
-    "build:lib": "node scripts/build-lib.js",
-    "build": "npm run clean && npm run build:docs && npm run build:code",
-    "pretest": "node scripts/pretest.js",
-    "test": "node scripts/test.js",
-    "test:watch": "react-app-rewired test",
-    "test:coverage": "react-app-rewired test --coverage --watchAll=false",
-    "release:patch": "lerna version patch --force-publish --yes",
-    "release:minor": "lerna version minor --force-publish --yes",
-    "release:major": "lerna publish major --force-publish --yes",
-    "alpha:patch": "lerna version prepatch --force-publish --yes",
-    "alpha:minor": "lerna version preminor --force-publish --yes",
-    "alpha:major": "lerna version premajor --force-publish --yes",
-    "clean": "node scripts/clean.js",
-    "docgen": "npm run docgen:examples && npm run docgen:props && npm run docgen:markdown",
-    "docgen:props": "node scripts/docgen-props.js",
-    "docgen:examples": "node scripts/docgen-examples.js",
-    "docgen:markdown": "node scripts/docgen-markdown.js",
-    "update-npmignore": "lerna exec -- cp \\$LERNA_ROOT_PATH/.npmignore .npmignore",
-    "deploy:docs": "npm run build:docsite && git add docs -f && git commit --allow-empty -m 'Deploying docs' && git push --force && npm run clean",
-    "changelog": "changelog || echo 'No Changes'",
-    "storybook": "start-storybook -p 6006",
-    "storybook:update": "npm i -D @storybook/addon-actions@latest @storybook/addon-knobs@latest @storybook/react@latest",
-    "preversion": "npm run clean && npm run docgen && npm run build:docsite && git add --all && git commit --allow-empty -m 'Generating Props'",
-    "version": "node scripts/fix-package-version && npm run changelog && git add CHANGELOG.md && git commit --allow-empty -m 'Updating Changelog'",
-    "perpare": "npm run build:code",
-    "build-storybook": "build-storybook"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.14.5",
-    "@storybook/addon-actions": "^6.2.9",
-    "@storybook/addon-essentials": "^6.2.9",
-    "@storybook/addon-knobs": "^6.2.9",
-    "@storybook/addon-links": "^6.2.9",
-    "@storybook/react": "^6.2.9",
-    "@testing-library/react": "^11.2.7",
-    "@types/enzyme": "^3.10.8",
-    "@types/enzyme-adapter-react-16": "^1.0.6",
-    "@types/jest": "^26.0.23",
-    "@types/node": "^15.12.2",
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.7",
-    "@types/react-router": "^5.1.15",
-    "@types/react-router-dom": "^5.1.7",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.2",
-    "babel-loader": "^8.2.2",
-    "colors": "^1.4.0",
-    "copyfiles": "^2.4.1",
-    "documentalist": "^1.6.0",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.6",
-    "eslint-plugin-babel": "^5.3.1",
-    "generate-changelog": "^1.8.0",
-    "lerna": "^3.22.1",
-    "module-alias": "^2.2.2",
-    "prettier": "2.3.1",
-    "react": "^17.0.2",
-    "react-app-rewired": "^2.1.8",
-    "react-dom": "^17.0.2",
-    "react-ga": "^3.3.0",
-    "react-live": "^2.2.3",
-    "react-markdown": "^6.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
-    "typescript": "4.2.2"
-  },
-  "eslintIgnore": [
-    "node_modules"
-  ],
-  "config-overrides-path": "config/rewire",
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+	"name": "rmwc-main",
+	"author": "James Friedman",
+	"version": "6.1.1",
+	"private": true,
+	"homepage": "https://rmwc.io",
+	"description": "RMWC is a React UI Kit built on Google's official Material Components Web library.",
+	"license": "MIT",
+	"sideEffects": true,
+	"main": "index.js",
+	"types": "index.d.ts",
+	"peerDependencies": {
+		"react": ">=16.8.x",
+		"react-dom": ">=16.8.x"
+	},
+	"scripts": {
+		"install": "lerna bootstrap --force-local --hoist",
+		"start": "export NODE_OPTIONS='--max-old-space-size=4096' && react-app-rewired start",
+		"build:docsite": "export NODE_OPTIONS='--max-old-space-size=4096' && react-app-rewired build && rm -R -f docs && mv build docs",
+		"build:docs": "npm run clean && npm run docgen && npm run build:docsite",
+		"build:code": "npm run clean && node scripts/build-code.js",
+		"build:lib": "node scripts/build-lib.js",
+		"build": "npm run clean && npm run build:docs && npm run build:code",
+		"pretest": "node scripts/pretest.js",
+		"test": "node scripts/test.js",
+		"test:watch": "react-app-rewired test",
+		"test:coverage": "react-app-rewired test --coverage --watchAll=false",
+		"release:patch": "lerna version patch --force-publish --yes",
+		"release:minor": "lerna version minor --force-publish --yes",
+		"release:major": "lerna publish major --force-publish --yes",
+		"alpha:patch": "lerna version prepatch --force-publish --yes",
+		"alpha:minor": "lerna version preminor --force-publish --yes",
+		"alpha:major": "lerna version premajor --force-publish --yes",
+		"clean": "node scripts/clean.js",
+		"docgen": "npm run docgen:examples && npm run docgen:props && npm run docgen:markdown",
+		"docgen:props": "node scripts/docgen-props.js",
+		"docgen:examples": "node scripts/docgen-examples.js",
+		"docgen:markdown": "node scripts/docgen-markdown.js",
+		"update-npmignore": "lerna exec -- cp \\$LERNA_ROOT_PATH/.npmignore .npmignore",
+		"deploy:docs": "npm run build:docsite && git add docs -f && git commit --allow-empty -m 'Deploying docs' && git push --force && npm run clean",
+		"changelog": "changelog || echo 'No Changes'",
+		"storybook": "start-storybook -p 6006",
+		"storybook:update": "npm i -D @storybook/addon-actions@latest @storybook/addon-knobs@latest @storybook/react@latest",
+		"preversion": "npm run clean && npm run docgen && npm run build:docsite && git add --all && git commit --allow-empty -m 'Generating Props'",
+		"version": "node scripts/fix-package-version && npm run changelog && git add CHANGELOG.md && git commit --allow-empty -m 'Updating Changelog'",
+		"prepare": "npm run build:code",
+		"build-storybook": "build-storybook"
+	},
+	"devDependencies": {
+		"@babel/core": "^7.17.9",
+		"@storybook/addon-actions": "^6.4.22",
+		"@storybook/addon-essentials": "^6.4.22",
+		"@storybook/addon-knobs": "^6.4.0",
+		"@storybook/addon-links": "^6.4.22",
+		"@storybook/builder-webpack5": "^6.4.22",
+		"@storybook/manager-webpack5": "^6.4.22",
+		"@storybook/react": "^6.4.22",
+		"@storybook/preset-create-react-app": "4.1.0",
+		"@testing-library/react": "^11.2.7",
+		"@types/enzyme": "^3.10.12",
+		"@types/enzyme-adapter-react-16": "^1.0.6",
+		"@types/jest": "^27.4.1",
+		"@types/node": "^15.12.2",
+		"@types/react": "^17.0.44",
+		"@types/react-dom": "^17.0.15",
+		"@types/react-router": "^5.1.18",
+		"@types/react-router-dom": "^5.1.7",
+		"@wojtekmaj/enzyme-adapter-react-17": "^0.6.2",
+		"babel-loader": "^8.2.5",
+		"colors": "^1.4.0",
+		"copyfiles": "^2.4.1",
+		"documentalist": "^1.6.0",
+		"enzyme": "^3.11.0",
+		"enzyme-adapter-react-16": "^1.15.6",
+		"eslint-plugin-babel": "^5.3.1",
+		"generate-changelog": "^1.8.0",
+		"lerna": "^3.22.1",
+		"module-alias": "^2.2.2",
+		"prettier": "2.3.1",
+		"react": "^17.0.2",
+		"react-app-rewired": "^2.2.1",
+		"react-dom": "^17.0.2",
+		"react-ga": "^3.3.0",
+		"react-live": "^3.1.0",
+		"react-markdown": "^6.0.2",
+		"react-router-dom": "^6.3.0",
+		"react-scripts": "^5.0.1",
+		"typescript": "4.3.5"
+	}
+	,
+	"eslintIgnore": [
+		"node_modules"
+	],
+	"config-overrides-path": "config/rewire",
+	"browserslist": [
+		">0.2%",
+		"not dead",
+		"not ie <= 11",
+		"not op_mini all"
+	]
 }
