@@ -5,19 +5,26 @@ import { useId, emptyClientRect } from '@rmwc/base';
 import { useFoundation } from '@rmwc/base';
 import { MDCChipFoundation, MDCChipAdapter } from '@material/chips';
 import { EventSource } from '@material/chips/chip/constants';
-import React, { useEffect, useCallback } from 'react';
+import React, { useEffect, useCallback, useRef } from 'react';
+import { TrailingActionApi } from './trailing-action';
 
 export const useChipFoundation = (props: ChipProps & ChipHTMLProps) => {
   const chipId = useId('chip', props);
+
+  const trailingAction = useRef<TrailingActionApi | null>();
+  const setTrailingAction = (api: TrailingActionApi | null) => {
+    trailingAction.current = api;
+  };
 
   const foundationWithElements = useFoundation({
     props,
     elements: {
       rootEl: true,
       trailingIconEl: true,
-      checkmarkEl: true
+      checkmarkEl: true,
+      trailingActionEl: true
     },
-    foundation: ({ rootEl, checkmarkEl, emit, getProps }) =>
+    foundation: ({ rootEl, checkmarkEl, emit, getProps, trailingActionEl }) =>
       new MDCChipFoundation({
         addClass: (className) => {
           rootEl.addClass(className);
@@ -77,16 +84,19 @@ export const useChipFoundation = (props: ChipProps & ChipHTMLProps) => {
         focusPrimaryAction: () => {
           // Not clear in documentation what this should be used for
         },
-        hasTrailingAction: () => {
-          return !!getProps().trailingIcon;
-        },
         setTrailingActionAttr: (attr: string, value: string) => {
           const safeAttr = attr === 'tabindex' ? 'tabIndex' : attr;
-          trailingIconEl.setProp(safeAttr as any, value);
+          return trailingActionEl.setProp(safeAttr as any, value);
         },
 
         focusTrailingAction: () => {
-          trailingIconEl.ref?.focus();
+          return trailingActionEl.ref?.focus();
+        },
+        removeTrailingActionFocus: () => {
+          return trailingAction.current?.getFoundation().removeFocus();
+        },
+        isTrailingActionNavigable: () => {
+          return trailingAction.current?.getFoundation().isNavigable();
         },
         isRTL: () => {
           return rootEl.ref
@@ -94,6 +104,12 @@ export const useChipFoundation = (props: ChipProps & ChipHTMLProps) => {
                 .getComputedStyle(rootEl.ref)
                 .getPropertyValue('direction') === 'rtl'
             : false;
+        },
+        notifyEditFinish: () => {
+          // NOT IMPLEMENTED IN MATERIAL 7
+        },
+        notifyEditStart: () => {
+          // NOT IMPLEMENTED IN MATERIAL 7
         }
       } as MDCChipAdapter)
   });
@@ -107,7 +123,7 @@ export const useChipFoundation = (props: ChipProps & ChipHTMLProps) => {
     ) => {
       evt.type === 'click' && onClick?.(evt as any);
       evt.type === 'keydown' && onKeyDown?.(evt as any);
-      return foundation.handleInteraction(evt);
+      return foundation.handleClick();
     },
     [foundation, onClick, onKeyDown]
   );
@@ -119,12 +135,9 @@ export const useChipFoundation = (props: ChipProps & ChipHTMLProps) => {
     [foundation]
   );
 
-  const handleTrailingIconInteraction = useCallback(
-    (evt: any) => {
-      return foundation.handleTrailingIconInteraction(evt);
-    },
-    [foundation]
-  );
+  const handleTrailingActionInteraction = useCallback(() => {
+    return foundation.handleTrailingActionInteraction();
+  }, [foundation]);
 
   // Allow customizing the behavior of the trailing icon
   useEffect(() => {
@@ -136,8 +149,8 @@ export const useChipFoundation = (props: ChipProps & ChipHTMLProps) => {
   rootEl.setProp('onClick', handleInteraction, true);
   rootEl.setProp('onKeyDown', handleInteraction, true);
   rootEl.setProp('onTransitionEnd', handleTransitionEnd, true);
-  trailingIconEl.setProp('onClick', handleTrailingIconInteraction, true);
-  trailingIconEl.setProp('onKeyDown', handleTrailingIconInteraction, true);
+  trailingIconEl.setProp('onClick', handleTrailingActionInteraction, true);
+  trailingIconEl.setProp('onKeyDown', handleTrailingActionInteraction, true);
 
-  return foundationWithElements;
+  return { setTrailingAction, ...foundationWithElements };
 };
