@@ -22,7 +22,7 @@ export default function Readme() {
   return (
     <Docs
       title="Grid Lists"
-      lead="Grid lists are an alternative to standard list views."
+      lead="Grid lists are an alternative to standard list views. Note that as of Material Components Web 6.0.0, this component is deprecated."
       module="@rmwc/grid-list"
       styles={['@material/grid-list/dist/mdc.grid-list.css']}
       docsLink="https://material.io/components/web/catalog/grid-lists/"
