@@ -84,27 +84,6 @@ const RESOURCES: ResourceT[] = [
     actions: [{ label: 'Start Chatting', url: 'https://discord.gg/4BSUxCW' }]
   },
   {
-    name: 'Donate on OpenCollective',
-    description:
-      'This library is built with Typescript, sweat, and thousands of hours of hard work. Help support Open Source Software!',
-    url: 'https://opencollective.com/rmwc',
-    img: 'images/backgrounds/mb-bg-fb-07.png',
-    icon: (
-      <svg viewBox="0 0 24 24">
-        <path
-          fill="inherit"
-          d="M22,12V20A2,2 0 0,1 20,22H4A2,2 0 0,1 2,20V12A1,1 0 0,1 1,11V8A2,2 0 0,1 3,6H6.17C6.06,5.69 6,5.35 6,5A3,3 0 0,1 9,2C10,2 10.88,2.5 11.43,3.24V3.23L12,4L12.57,3.23V3.24C13.12,2.5 14,2 15,2A3,3 0 0,1 18,5C18,5.35 17.94,5.69 17.83,6H21A2,2 0 0,1 23,8V11A1,1 0 0,1 22,12M4,20H11V12H4V20M20,20V12H13V20H20M9,4A1,1 0 0,0 8,5A1,1 0 0,0 9,6A1,1 0 0,0 10,5A1,1 0 0,0 9,4M15,4A1,1 0 0,0 14,5A1,1 0 0,0 15,6A1,1 0 0,0 16,5A1,1 0 0,0 15,4M3,8V10H11V8H3M13,8V10H21V8H13Z"
-        />
-      </svg>
-    ),
-    actions: [
-      {
-        label: 'Donate',
-        url: 'https://opencollective.com/rmwc'
-      }
-    ]
-  },
-  {
     name: 'Share',
     description:
       "Don't keep letting your friends build software with bad UI libraries. Spread the word about RMWC.",
@@ -193,8 +172,7 @@ const RESOURCES: ResourceT[] = [
     actions: [
       {
         label: 'Read Guidelines',
-        url:
-          'https://github.com/jamesmfriedman/rmwc/blob/master/CONTRIBUTING.md'
+        url: 'https://github.com/jamesmfriedman/rmwc/blob/master/CONTRIBUTING.md'
       }
     ]
   }
@@ -206,7 +184,7 @@ export default function Resources() {
       <h1>Resources</h1>
       <div className={styles.cardGrid}>
         <div className={styles.cardGridInner}>
-          {RESOURCES.map(r => (
+          {RESOURCES.map((r) => (
             <ResourceCard key={r.name} resource={r} />
           ))}
         </div>
@@ -245,7 +223,7 @@ function ResourceCard({ resource }: { resource: ResourceT }) {
 
       <CardActions>
         <CardActionButtons>
-          {resource.actions.map(a => (
+          {resource.actions.map((a) => (
             <CardActionButton
               key={a.label}
               tag="a"
