@@ -19,6 +19,7 @@ import { withRipple } from '@rmwc/ripple';
 
 import { useSelectFoundation } from './foundation';
 import { SelectIcon } from '../select-icon';
+import { Icon } from '@rmwc/icon';
 
 export interface FormattedOption
   extends Omit<React.AllHTMLAttributes<any>, 'label'> {
@@ -95,9 +96,17 @@ const createSelectOptions = (options: any): FormattedOption[] => {
   return options;
 };
 
-const SelectDropdownArrow = React.memo(function SelectDropdownArrow() {
-  return <i className="mdc-select__dropdown-icon" />;
-});
+const SelectDropdownArrow = () => {
+  return (
+    <span className="mdc-select__dropdown-icon">
+      <Icon
+        className="mdc-select__dropdown-icon-inactive"
+        icon="arrow_drop_down"
+      />
+      <Icon className="mdc-select__dropdown-icon-active" icon="arrow_drop_up" />
+    </span>
+  );
+};
 
 function NativeMenu(
   props: {
@@ -114,6 +123,7 @@ function NativeMenu(
     placeholder = '',
     children,
     elementRef,
+    open,
     ...rest
   } = props;
 
@@ -176,7 +186,7 @@ function NativeMenu(
   );
 }
 
-const SelectedTextEl = withRipple({ surface: false })(function (props: any) {
+const AnchorEl = withRipple({ surface: false })(function (props: any) {
   return <Tag {...props} />;
 });
 
@@ -203,6 +213,8 @@ function EnhancedMenu(props: EnhancedMenuProps & SelectHTMLProps) {
 
   let currentIndex = 0;
 
+  const className = useClassNames(props, ['mdc-select__menu']);
+
   const renderOption = ({
     label,
     option
@@ -223,25 +235,20 @@ function EnhancedMenu(props: EnhancedMenuProps & SelectHTMLProps) {
         {...option}
         data-value={option.value}
       >
-        {label}
+        <span className="mdc-list-item__text">{label}</span>
       </MenuItem>
     );
   };
 
   return (
-    <Menu
-      {...rest}
-      apiRef={menuApiRef}
-      className="mdc-select__menu"
-      focusOnOpen
-    >
+    <Menu {...rest} apiRef={menuApiRef} className={className} focusOnOpen>
       {!!props.placeholder && (
         <MenuItem
           selected={currentIndex - 1 === selectedIndex}
           data-value=""
           theme="textDisabledOnBackground"
         >
-          {placeholder}
+          <span className="mdc-list-item__text">{placeholder}</span>
         </MenuItem>
       )}
 
@@ -302,7 +309,7 @@ export const Select: RMWC.ComponentType<
   const selectOptions = createSelectOptions(options);
   const {
     rootEl,
-    selectedTextEl,
+    anchorEl,
     notchWidth,
     menuOpen,
     selectedTextContent,
@@ -317,6 +324,7 @@ export const Select: RMWC.ComponentType<
     handleFocus,
     handleBlur,
     handleClick,
+    handleChange,
     handleKeydown,
     handleMenuClosed,
     handleMenuOpened,
@@ -329,6 +337,7 @@ export const Select: RMWC.ComponentType<
     'mdc-select',
     {
       'mdc-select--outlined': !!outlined,
+      'mdc-select--filled': !outlined,
       'mdc-select--required': !!props.required,
       'mdc-select--invalid': !!invalid,
       'mdc-select--with-leading-icon': !!icon,
@@ -366,38 +375,35 @@ export const Select: RMWC.ComponentType<
 
   return (
     <>
-      <Tag
-        role="listbox"
-        {...rootProps}
-        element={rootEl}
-        ref={ref}
-        className={className}
-      >
-        <div className="mdc-select__anchor">
+      <Tag {...rootProps} element={rootEl} ref={ref} className={className}>
+        <AnchorEl
+          className="mdc-select__anchor"
+          role="button"
+          aria-haspopup="listbox"
+          element={anchorEl}
+          onFocus={handleFocus}
+          onBlur={handleBlur}
+          onClick={handleClick}
+          onKeyDown={handleKeydown}
+          onChange={handleChange}
+          /** In the case of native selects, we don't want this to be be focusable */
+          tabIndex={enhanced ? undefined : -1}
+        >
           {!!icon && <SelectIcon apiRef={setLeadingIcon} icon={icon} />}
-          <SelectDropdownArrow />
-          <SelectedTextEl
-            className="mdc-select__selected-text"
-            role="button"
-            aria-haspopup="listbox"
-            element={selectedTextEl}
-            onFocus={handleFocus}
-            onBlur={handleBlur}
-            onClick={handleClick}
-            onKeyDown={handleKeydown}
-            /** In the case of native selects, we don't want this to be be focusable */
-            tabIndex={enhanced ? undefined : -1}
-          >
-            {selectedTextContent || <>&nbsp;</>}
-          </SelectedTextEl>
           {outlined ? (
             <NotchedOutline notch={notchWidth}>{renderedLabel}</NotchedOutline>
           ) : (
             <>
-              {renderedLabel}
               <LineRipple active={lineRippleActive} center={lineRippleCenter} />
+              {renderedLabel}
             </>
           )}
+          <span className="mdc-select__selected-text-container">
+            <span className="mdc-select__selected-text">
+              {selectedTextContent}
+            </span>
+          </span>
+          <SelectDropdownArrow />
           {!enhanced && (
             <NativeMenu
               {...rest}
@@ -408,13 +414,19 @@ export const Select: RMWC.ComponentType<
               selectOptions={selectOptions}
               elementRef={setNativeControl}
               onFocus={handleFocus}
-              onBlur={handleBlur}
-              onChange={(evt: React.ChangeEvent<HTMLSelectElement>) =>
-                handleMenuSelected(evt.currentTarget.selectedIndex)
-              }
+              onBlur={(evt) => {
+                handleBlur(evt);
+                // As of material 7, there is a bug where classnames mdc-select--activated and mdc-select--focused
+                // isn't being removed on blur. Therefore we need to manually call handleMenuClosed onBlur, which
+                // takes care of removing the classnames.
+                handleMenuClosed();
+              }}
+              onChange={(evt: React.ChangeEvent<HTMLSelectElement>) => {
+                handleMenuSelected(evt.currentTarget.selectedIndex);
+              }}
             />
           )}
-        </div>
+        </AnchorEl>
 
         {enhanced && (
           <EnhancedMenu
