@@ -31,9 +31,6 @@ export default function Readme() {
       <DocsExample label="Outlined">
         <TextField outlined label="outlined..." />
       </DocsExample>
-      <DocsExample label="Full Width">
-        <TextField fullwidth placeholder="fullWidth..." />
-      </DocsExample>
       <DocsExample label="No Label">
         <TextField placeholder="No label" />
       </DocsExample>
@@ -62,20 +59,21 @@ export default function Readme() {
         maxLength property.
       </DocsP>
       <DocsExample>
-        <TextField
-          textarea
-          outlined
-          fullwidth
-          label="textarea..."
-          rows={8}
-          maxLength={20}
-          characterCount
-          helpText={{
-            persistent: true,
-            validationMsg: true,
-            children: 'The field is required'
-          }}
-        />
+        <div>
+          <TextField
+            textarea
+            label="textarea..."
+            rows={8}
+            maxLength={20}
+            characterCount
+            resizeable
+            helpText={{
+              persistent: true,
+              validationMsg: true,
+              children: 'The field is required'
+            }}
+          />
+        </div>
       </DocsExample>
 
       <DocsSubtitle>Validation</DocsSubtitle>
